Split lsp.rs and query.rs into subcommand + adapter modules
Separate concerns so adapters can grow independently:
- lsp.rs: subcommand entry point (main loop, message dispatch)
- lsp_adapter.rs: LSP JSON-RPC transport + diagnostics formatting
- query.rs: subcommand entry point (arg parsing, dispatch)
- query_adapter.rs: Engine invocation + output formatting

Also made transport functions generic over AsyncRead/AsyncWrite
instead of concrete tokio::io::Stdin/Stdout.
