+++ wado-cli/src/lib.rs
@@ -20,8 +20,10 @@ pub mod dump;
 pub mod format;
 pub mod init;
 pub mod lsp;
+pub mod lsp_adapter;
 pub mod manifest;
 pub mod query;
+pub mod query_adapter;
 pub mod run;
 pub mod runtime;
 pub mod serve;
